+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
-## [1.0.0] - YYYY-MM-DD
+## [0.1.0] - 2026-03-12
 
-Initial release.
+### Added
+
+- `SpriteTKBundleBuilder` which can be used to build a sprite sheet bundle.
+- `SpriteTKBundleLoader` which can be used to load a sprite sheets bundle.